demo/requirements.txt
pandas==1.5.3
Pillow==10.3.0
Requests==2.31.0
streamlit==1.32.2
streamlit-authenticator==0.3.2
python-dotenv==1.0.1
matplotlib==3.9.0